% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_allometric.R
\name{get_allometric}
\alias{get_allometric}
\title{Get allometric data from FishBase for a single species}
\usage{
get_allometric(genus, species)
}
\arguments{
\item{genus}{Name of the genus of the fish (ex: "Brycon")}

\item{species}{Name of the fish species (ex: "orthotaenia")}
}
\value{
A \code{data.frame} with the allometric coefficients \code{a}, \code{b} and other coefficients.
}
\description{
This function web scrapes the FishBase (https://fishbase.se/) website to obtain the coefficients of the length-weight equation.
}
\details{
Agronomist | Biodiversity - MS | Soil Science - PhD
}
\seealso{
\url{mailto:rzaratesalazar@gmail.com}
}
\author{
Zárate-Salazar, Jhonatan Rafael (Researcher | PPEC - UFS)
}
