% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fish_allometric.R
\name{get_fish_allometric}
\alias{get_fish_allometric}
\title{Get allometric data from FishBase for multiple species}
\usage{
get_fish_allometric(genus, species)
}
\arguments{
\item{genus}{character vector with of the name of the genus of the fish}

\item{species}{character verctor with of the name of the fish species}
}
\value{
A \code{data.frame} with the allometric coefficients \code{a}, \code{b} and other coefficients.
}
\description{
This function web scrapes the FishBase (https://fishbase.se/) website to obtain the coefficients of the length-weight equation.
}
\details{
Agronomist | Biodiversity - MS | Soil Science - PhD
}
\seealso{
\url{mailto:rzaratesalazar@gmail.com}
}
\author{
Zárate-Salazar, Jhonatan Rafael (Researcher | PPEC - UFS)
}
